Add hr module and dbc class
