+++ student.rb
@@ -1,16 +1,11 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'dbc'
+
+class Student < Dbc
 
   def initialize(options = {})
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
 
   def set_phase(num)
     response = ""
@@ -20,8 +15,7 @@ class Student
     else
       response = "Oooh, phase #{num}. I hope I'm ready!"
     end
-    @phase = num
-    response
+    super
   end
 
   def learn_stuff
