+++ teacher.rb
@@ -1,22 +1,17 @@
-class Teacher
-  attr_reader :age, :salary, :phase, :performance_rating, :target_raise
+require_relative 'dbc'
+require_relative 'hr'
+class Teacher < Dbc
+  include Hr
+
+  attr_reader :salary, :performance_rating, :target_raise
   attr_accessor :name
 
   def initialize(options={})
     @phase = 3
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+
     @target_raise = 1000
   end
 
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
 
   def teach_stuff
     response = ""
@@ -26,24 +21,12 @@ class Teacher
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
 
-  def receive_raise(raise)
-    @salary += raise
-  end
 
   def set_performance_rating(rating)
     response = ""
     if rating > 90
-      receive_raise(@target_raise)
-      response = "Yay, I'm a great employee!"
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
+      super
     end
-    response
   end
 end
